+++ unittests/utils.py
@@ -107,17 +107,21 @@ def polar_gravity_acceleration(bs, component=None, components_distance=None):
 
 
 def prepare_binary_system(params, spots_primary=None, spots_secondary=None):
+    pkwargs = dict(discretization_factor=params.get("primary_discretization_factor")) \
+        if "primary_discretization_factor" in params else dict()
     primary = Star(mass=params["primary_mass"], surface_potential=params["primary_surface_potential"],
                    synchronicity=params["primary_synchronicity"],
                    t_eff=params["primary_t_eff"], gravity_darkening=params["primary_gravity_darkening"],
                    albedo=params['primary_albedo'], metallicity=0.0, spots=spots_primary,
-                   discretization_factor=params.get("primary_discretization_factor", 3))
+                   **pkwargs)
 
+    skwargs = dict(discretization_factor=params.get("secondary_discretization_factor")) \
+        if "secondary_discretization_factor" in params else dict()
     secondary = Star(mass=params["secondary_mass"], surface_potential=params["secondary_surface_potential"],
                      synchronicity=params["secondary_synchronicity"],
                      t_eff=params["secondary_t_eff"], gravity_darkening=params["secondary_gravity_darkening"],
                      albedo=params['secondary_albedo'],
-                     metallicity=0.0, spots=spots_secondary)
+                     metallicity=0.0, spots=spots_secondary, **skwargs)
 
     return BinarySystem(primary=primary,
                         secondary=secondary,